register page now redirects to dashboard
